Fixing Jitrino Linux build regression introduced by warning removal on Win64 platform.
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@568956 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/jitrino/src/optimizer/loop_unroll.cpp
@@ -237,7 +237,7 @@ void LoopUnrollPass::_run(IRManager& irm) {
 
         //check config settings
         bool failed = false;
-        size_t nodesInLoop = loopHeader->getNodesInLoop().size();
+        int nodesInLoop = (int)loopHeader->getNodesInLoop().size();
         const char* reason = "unknown";
         if (nodesInLoop > flags.largeLoopSize) {
             reason = "loop is too large";
